+++ lib/widgets/widgets.dart
@@ -1,0 +1,2 @@
+export 'animated_digit_widget.dart';
+export 'digit_widget.dart';